pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "miniflux-summary-agent"
version = "0.1.0"
description = "RSS article summarization agent using Miniflux and Ollama AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=0.19.0",
    "jinja2>=3.1.6",
    "miniflux>=1.1.4",
    "ollama>=0.5.1",
    "playwright>=1.53.0",
    "pydantic>=2.0.0",
    "requests>=2.32.4",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
miniflux-summary = "miniflux_summary_agent.cli:cli_main"

[tool.setuptools.packages.find]
include = ["miniflux_summary_agent*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"miniflux_summary_agent" = ["templates/*.html"]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing --cov-report=html --cov-report=xml"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    ".venv/*",
    "__pycache__/*",
    "tests/*",
    "*/.git/*",
    ".pytest_cache/*",
    "htmlcov/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
show_missing = true

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["miniflux_summary_agent"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
